Fix '//' issue with cygwin $prefix paths.
Normally the configure --prefix setting is an absolute path.  Two of the genders extension
Makefiles (perl/Genders and python) had an extra slash in front of $prefix.

On Unix systems is is benign; paths that start with '//' are equivalent to paths
that start with '/'.

But on cygwin / ms-windows systems '//' is 'magic', and specified a global network path as
opposed to a host relative path.

I deleted the extra slash in front of $prefix and now 'make install' runs to completion
on cygwin.