Build Bindings for *both* python2 and python3
Update the spec file to do multiple builds (if necessary) for *both*
python2 and python3 bindings.  Both get built by default, but can be
controlled with _with_python2 or _with_python3.  Continue to respect
the _with_python_extensions macro as well.  Currently tested on a
TOSS3 host, and should work on a TOSS4 host (but still needs to be
tested).  Should be able to drop the genders.t4.patch with this update.