libgenders: remove duplicate code
Problem: In prior commit

ba9a7d9ecc57bcac1412938cb6d4295c5b8057db

a parsing corner case was handled by copying some code and
modifying it.  However, the original code was not deleted, leading
to a double check on some values.  This is wasted cycles.

Remove the extra code that is unnecessary.