libgenders: improve duplicate attrs check
Problem: Checking for duplicate attrs on the same line of a genders
file is done very inefficiently.  It is done for every node specified
on the line.  It would be sufficient to check this list only once
regardless of how many nodes are on the line.

Remove duplicate attr checks on the same line from _attr_node_processing()
to _insert_attrval().  This ensures that duplicates on the same line
are only checked once per line.