libgenders: increase default hash sizes
Problem: Clusters are much larger nowadays and the default hash
sizes are pretty small.  It leads to internal re-hashing of the
hashes.  The minor extra memory cost for larger hashes is well worth
it for extra speed up on larger clusters.

Increase the default hash size for many internal data structures.