libgenders: set temp variable before return value
Problem: A call to _attr_node_processing() sets the "rv" return
value for the function _parse_line().  This happens to work because
this is the last function call before the end of the function.  But
new code could be added in the future.

Have the return value from _attr_node_processing() set a temp variable
instead.  Only set the "rv" return value on error before going to
the cleanup path.